requirements.txt
apikit==0.3.2
asn1crypto==0.24.0
bcrypt==3.1.4
CacheControl==0.12.3
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
cssmin==0.2.0
decorator==4.1.2
enum34==1.1.6
Fabric==1.14.0
Flask==0.12.2
Flask-Assets==0.12
Flask-Cache==0.13.1
Flask-FlatPages==0.6
Flask-Script==2.0.6
# grano-client==0.4.5
# granoloader==0.4.5
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
lockfile==0.12.2
lxml==4.1.1
Markdown==2.6.10
MarkupSafe==1.0
msgpack-python==0.4.8
networkx==2.0
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.6.1
python-json-logger==0.1.8
python-Levenshtein==0.12.0
python-slugify==1.2.4
PyYAML==3.12
requests==2.18.4
scrapekit==0.2.1
six==1.11.0
-e git+https://github.com/codeforafrica/connectedAFRICA.git@13a8eb4e1102b50803372ad1c4bfbf225efa10b7#egg=Siyazana
thready==0.1.5
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
webassets==0.12.1
Werkzeug==0.13
file:///app/repo/grano-client
file:///app/repo/granoloader
